:bug:Fix wrong type

+++ src/main/scala/com/keivanabdi/scalajs/epochjs/chart/basic/line/Value.scala
@@ -1,13 +1,13 @@
 package com.keivanabdi.scalajs.epochjs.chart.basic.line
 
 class Value(
-    val x: Int,
+    val x: Double,
     val y: Double
 ) extends scalajs.js.Object
 
 object Value {
   def apply(
-      x: Int,
+      x: Double,
       y: Double
   ): Value = new Value(x, y)
 }
